pyproject.toml
[project]
name = "devmate"
version = "0.1.0"
description = "DevMate intelligent coding agent for developer workflows"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "langchain>=1.2.10",
    "langchain-core>=1.2.23",
    "langchain-openai>=0.3.0",
    "langsmith>=0.3.0",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-text-splitters>=0.3.8",
    "langchain-chroma>=0.2.6",
    "mcp==1.24.0",
    "httpx==0.28.1",
    "uvicorn==0.42.0",
    "starlette>=0.46.0",
    "tavily-python>=0.5.0",
    "chromadb>=1.0.0",
    "langchain-community>=0.3.0"
]

[project.scripts]
devmate = "devmate.main:main"
devmate-test = "devmate.scripts.test:main"
devmate-run-mcp-server = "devmate.scripts.run_mcp_server:main"
devmate-test-mcp-client = "devmate.scripts.test_mcp_client:main"
devmate-build-kb = "devmate.scripts.build_kb:main"
devmate-test-rag = "devmate.scripts.test_rag:main"



[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["devmate"]


